Warning when s.t. uses FFI but has no FFI file, or vice versa.
This could be done in a much easier fashion by simply checking for FFI
declarations straight away in `convert`. However, I want to later be
able to perform other checks at later stages of the conversion process,
so I'm setting up some infrastructure for that as well. Specifically,
when checking whether strings contain interpolataions, we want to take
not of that so we can emit warnings later, but that's coming in a
separate PR.