Convert: Use Haskell's string handling
We previously used `decodeString` to convert from PSString to Text.
`decodeString`, and the rest of PS's string handling, take extra care to
make sure it doesn't output invalid unicode code points. That's nice,
but it was also the only place where we ever throw an error during
conversion, so it's not really worth it. Instead, we now just use
Haskell's default mechanisms for string conversion.